Delay timbre library load to fix initialization bug
Fixes tgwizard/bpm-game#3

The timbre library attemps to start an AudioContext on load.
This fails in newer versions of Chrome because autoplay is blocked
until the user has made an interaction.
The best fix is for the timbre library to wait for interaction
or call context.resume() before attempting to play audio.
However, the timbre repo is no longer maintained.
As a patch fix, dynamically load the library after the first
interaction in the guess tempo game (READY is clicked).
The timbre library is removed entirely from the keep bpm game
as it is not used.